fix: Change bnb_4bit_compute_dtype to float16 to support Kaggle T4 GPUs

+++ src/trainer.py
@@ -21,7 +21,7 @@ def train_model():
     bnb_config = BitsAndBytesConfig(
         load_in_4bit=True,
         bnb_4bit_quant_type="nf4",
-        bnb_4bit_compute_dtype=torch.bfloat16,
+        bnb_4bit_compute_dtype=torch.float16,
         bnb_4bit_use_double_quant=True,
     )
